% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{anonymize_content}
\alias{anonymize_content}
\title{Anonymizes personal data in a text}
\usage{
anonymize_content(
  con,
  content,
  positions,
  contacts,
  crm_lead_ids,
  all_participants,
  preserved_names,
  scope,
  content_id
)
}
\arguments{
\item{con}{Database connection}

\item{content}{The text to be anonymized}

\item{positions}{CRM positions data}

\item{contacts}{Contacts data}

\item{crm_lead_ids}{CRM lead IDs}

\item{all_participants}{All participants data}

\item{preserved_names}{A vector of names to be preserved}

\item{scope}{Scope of anonymization ("call_transcripts" or "crm_protocols")}

\item{content_id}{ID of the content being anonymized}
}
\value{
Text with anonymized personal data
}
\description{
Anonymizes personal data in a text
}
\keyword{internal}
