% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{anonymize_transcripts}
\alias{anonymize_transcripts}
\title{Anonymizes contents of the transcript table}
\usage{
anonymize_transcripts(
  con,
  call_participants,
  contacts,
  preserved_names,
  positions,
  lead_mails,
  complete_overwrite = FALSE
)
}
\arguments{
\item{con}{Database connection}

\item{call_participants}{A data frame of participants (raw.msgraph_call_participants)}

\item{contacts}{A data frame of contacts (raw.msgraph_contacts)}

\item{preserved_names}{A vector of names to preserve}

\item{positions}{A data frame of CRM positions}

\item{lead_mails}{A data frame of CRM lead mail addresses}

\item{complete_overwrite}{Logical, whether to overwrite existing anonymizations}
}
\value{
Invisible NULL
}
\description{
Anonymizes contents of the transcript table
}
\keyword{internal}
