% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{anonymize_with_crm_data}
\alias{anonymize_with_crm_data}
\title{Anonymizes customers in a token dataframe}
\usage{
anonymize_with_crm_data(con, crm_lead_ids, positions, tokens)
}
\arguments{
\item{con}{Database connection}

\item{crm_lead_ids}{A dataframe with CRM lead IDs}

\item{positions}{A dataframe with CRM positions}

\item{tokens}{A dataframe with tokens}
}
\value{
A dataframe with anonymized tokens for customers
}
\description{
Anonymizes customers in a token dataframe
}
\keyword{internal}
