% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{clean_phone_numbers}
\alias{clean_phone_numbers}
\title{Finds phone numbers in a text and removes spaces}
\usage{
clean_phone_numbers(text_vector)
}
\arguments{
\item{text_vector}{A character vector with text}
}
\value{
A character vector with cleaned phone numbers
}
\description{
Finds phone numbers in a text and removes spaces
}
\keyword{internal}
