% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{clean_token}
\alias{clean_token}
\title{Cleans tokens, filters out tokens in PERSON that are too short
Finds and replaces emails, phone numbers and websites}
\usage{
clean_token(tokens, preserved_names)
}
\arguments{
\item{tokens}{A dataframe with tokens}

\item{preserved_names}{A vector of names to be preserved}
}
\value{
A dataframe with cleaned tokens
}
\description{
Cleans tokens, filters out tokens in PERSON that are too short
Finds and replaces emails, phone numbers and websites
}
\keyword{internal}
