% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{compress_column}
\alias{compress_column}
\title{Compresses and normalizes company names}
\usage{
compress_column(column)
}
\arguments{
\item{column}{A character vector with company names}
}
\value{
A character vector with compressed company names
}
\description{
Compresses and normalizes company names
}
\keyword{internal}
