% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_lead_names.R
\name{create_changelog}
\alias{create_changelog}
\title{create_changelog}
\usage{
create_changelog(missing_activities, daily_download)
}
\arguments{
\item{missing_activities}{input_argument_description}
}
\value{
return_value_description
}
\description{
Create a changelog for Ansprechpartner changes based on the change of the first_name
of the Ansprechpartner in CRM from CRM/changelogs/missing_activities data.
This function also creates a second table, that tracks changes in either only the
first oder the last name.
}
\details{
Afterwards the newly created tables are saved tp ../base-data/CRM/changelogs/ folder
as RDS files.
}
