% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_main.R
\name{create_transcript_logger}
\alias{create_transcript_logger}
\title{Create Logger Function}
\usage{
create_transcript_logger(level = "INFO", file = NULL)
}
\description{
Create Logger Function
}
\keyword{internal}
