% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_crm_export.R
\name{crm_extract_protocol_id}
\alias{crm_extract_protocol_id}
\title{Read the protocol ID out of a CRM create response}
\usage{
crm_extract_protocol_id(response)
}
\arguments{
\item{response}{httr response object of the POST}
}
\value{
Protocol ID or NULL
}
\description{
Read the protocol ID out of a CRM create response
}
\keyword{internal}
