% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_lead_mails_background.R
\name{crm_update_lead_mails_background}
\alias{crm_update_lead_mails_background}
\title{Update CRM Lead Mails in Background}
\usage{
crm_update_lead_mails_background(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
No return value. Updates relevant tables in the database.
}
\description{
This function processes CRM lead email addresses, extracting, cleaning, and matching them with MS Graph contacts.
}
\examples{
crm_update_lead_mails_background(con)
}
