% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnectorR-package.R
\docType{package}
\name{dbconnectorR-package}
\alias{dbconnectorR}
\alias{dbconnectorR-package}
\title{dbconnectorR: Connect and Export Data from External Systems to Your Database}
\description{
Provides functions to extract, transform, and load data from external systems such as CRM platforms and Microsoft Graph into your internal database. Designed for streamlined data pipelines, standardized exports, and consistent data management.
}
\author{
\strong{Maintainer}: Moritz Hemmann \email{moritz.hemmann@studyflix.de}

}
\keyword{internal}
