% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_deanonymization.R
\name{deanonymize_text}
\alias{deanonymize_text}
\title{Deanonymizes text by replacing anonymized IDs with original names}
\usage{
deanonymize_text(con, anonymized_text, transcript_id_var)
}
\arguments{
\item{con}{Database connection}

\item{anonymized_text}{A character vector with anonymized text}

\item{transcript_id_var}{Transcript ID to look up placeholders for}
}
\value{
A character vector with deanonymized text
}
\description{
Deanonymizes text by replacing anonymized IDs with original names
}
\keyword{internal}
