% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_deanonymization.R
\name{deanonymize_transcript_summaries}
\alias{deanonymize_transcript_summaries}
\title{Adds deanonymized summaries to the transcript table}
\usage{
deanonymize_transcript_summaries(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
NULL (updates DB in place)
}
\description{
Adds deanonymized summaries to the transcript table
}
\keyword{internal}
