% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_main.R
\name{determine_transcript_resume_steps}
\alias{determine_transcript_resume_steps}
\title{Determine Which Steps to Resume}
\usage{
determine_transcript_resume_steps(con, logger)
}
\description{
Determine Which Steps to Resume
}
\keyword{internal}
