% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_main.R
\name{execute_transcript_anonymization}
\alias{execute_transcript_anonymization}
\title{Execute Transcript Anonymization Step}
\usage{
execute_transcript_anonymization(con, config)
}
\description{
Execute Transcript Anonymization Step
}
\keyword{internal}
