% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_main.R
\name{execute_transcript_crm_export}
\alias{execute_transcript_crm_export}
\title{Execute CRM Export Step}
\usage{
execute_transcript_crm_export(con, config)
}
\description{
Execute CRM Export Step
}
\keyword{internal}
