% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_main.R
\name{execute_transcript_load}
\alias{execute_transcript_load}
\title{Execute Transcript Load Step}
\usage{
execute_transcript_load(con, config)
}
\description{
Execute Transcript Load Step
}
\keyword{internal}
