% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_main.R
\name{execute_transcript_summarization}
\alias{execute_transcript_summarization}
\title{Execute Transcript Summarization Step}
\usage{
execute_transcript_summarization(con, config)
}
\description{
Execute Transcript Summarization Step
}
\keyword{internal}
