% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_crm_export.R
\name{export_all_transcripts_to_crm}
\alias{export_all_transcripts_to_crm}
\title{Export All Transcripts to CRM}
\usage{
export_all_transcripts_to_crm(
  con,
  crm_keys,
  use_test_account = FALSE,
  logger = function(msg, level = "INFO") cat(msg, "\\n")
)
}
\arguments{
\item{con}{Database connection}

\item{crm_keys}{CRM API key}

\item{use_test_account}{Logical, whether to use test CRM account}

\item{logger}{Logger function}
}
\value{
Invisible NULL
}
\description{
Exports all deanonymized transcript summaries to CRM as protocols
}
\keyword{internal}
