% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_scoped_events.R
\name{extract_meeting_id_safe}
\alias{extract_meeting_id_safe}
\title{Meeting-ID-Extraktion, NA-sicher}
\usage{
extract_meeting_id_safe(url)
}
\arguments{
\item{url}{Teams/OnlineMeeting joinUrl.}
}
\value{
meeting_id als character, oder NA_character_ bei Fehler.
}
\description{
Meeting-ID-Extraktion, NA-sicher
}
\keyword{internal}
