% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{filter_person_data_all}
\alias{filter_person_data_all}
\title{Finds personal data in a text and replaces it with a placeholder}
\usage{
filter_person_data_all(con, tokens_df, person_ids, company_ids)
}
\arguments{
\item{con}{Database connection}

\item{tokens_df}{A dataframe with tokens}

\item{person_ids}{The ids of the persons}

\item{company_ids}{The ids of the companies}
}
\value{
A dataframe with replaced personal data
}
\description{
Finds personal data in a text and replaces it with a placeholder
}
\keyword{internal}
