% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{filter_salesperson_crm_data}
\alias{filter_salesperson_crm_data}
\title{Finds salesperson data in CRM and returns tokens}
\usage{
filter_salesperson_crm_data(con, tokens_df, person_ids)
}
\arguments{
\item{con}{Database connection}

\item{tokens_df}{A dataframe with tokens}

\item{person_ids}{The ids of the persons}
}
\value{
A dataframe with salesperson tokens
}
\description{
Finds salesperson data in CRM and returns tokens
}
\keyword{internal}
