% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{find_and_replace_personal_data}
\alias{find_and_replace_personal_data}
\title{Finds a token and replaces it with a placeholder in the NER column}
\usage{
find_and_replace_personal_data(tokens_df, personal_data, replacement)
}
\arguments{
\item{tokens_df}{A dataframe with tokens}

\item{personal_data}{A character vector with personal data}

\item{replacement}{A character vector with replacements}
}
\value{
A dataframe with replaced personal data
}
\description{
Finds a token and replaces it with a placeholder in the NER column
}
\keyword{internal}
