% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{get_and_save_transcript_data}
\alias{get_and_save_transcript_data}
\title{Get and Save Transcript Data}
\usage{
get_and_save_transcript_data(
  con,
  authentication_msgraph,
  users,
  call_event_mapping,
  calls,
  start_date = NULL,
  end_date = Sys.Date() + 1
)
}
\description{
Main orchestrator for transcript retrieval process
}
\keyword{internal}
