% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_mails.R
\name{get_bidi_pattern}
\alias{get_bidi_pattern}
\title{Remove Bidirectional Unicode Characters}
\usage{
get_bidi_pattern()
}
\value{
Regular expression pattern for bidirectional characters
}
\description{
Returns pattern for problematic bidirectional Unicode characters.
}
\keyword{internal}
