% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{get_call_id}
\alias{get_call_id}
\title{Get Call ID from Meeting ID}
\usage{
get_call_id(metadata_meeting_id, creation_datetime_transcript, calls)
}
\description{
Matches transcript to call record
}
\keyword{internal}
