% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msgraph_func.R
\name{get_call_participants_combined}
\alias{get_call_participants_combined}
\title{Get Combined Call Participants}
\usage{
get_call_participants_combined(con, call_ids, include_event_attendees = TRUE)
}
\arguments{
\item{con}{Database connection}

\item{call_ids}{Vector of call IDs}

\item{include_event_attendees}{Logical, include event invitees}
}
\value{
Data frame with participant information
}
\description{
Retrieves all participants for calls, combining actual participants
with event invitees
}
\keyword{internal}
