% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{get_latest_transcript_timestamp}
\alias{get_latest_transcript_timestamp}
\title{Get Latest Transcript Timestamp}
\usage{
get_latest_transcript_timestamp(con)
}
\description{
Returns the latest transcript date from database
}
\keyword{internal}
