% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{get_meeting_metadata}
\alias{get_meeting_metadata}
\title{Get Meeting Metadata}
\usage{
get_meeting_metadata(access_token, user_id, meeting_id)
}
\description{
Gets full meeting metadata from MS Graph API
}
\keyword{internal}
