% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_mails.R
\name{get_processable_users}
\alias{get_processable_users}
\title{Get Processable Users}
\usage{
get_processable_users(con)
}
\arguments{
\item{con}{Database connection object (pool or DBI connection)}
}
\value{
Data frame with user information
}
\description{
Retrieves all internal, non-deleted users from the database.
}
