% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{get_transcript_row}
\alias{get_transcript_row}
\title{Get Transcript Row}
\usage{
get_transcript_row(
  entry,
  meeting_metadata,
  all_events_categorised_df,
  calls,
  access_token,
  user_id,
  existing_transcript_ids = character(0)
)
}
\description{
Processes a single transcript entry and returns formatted data
}
\keyword{internal}
