% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_deanonymization.R
\name{get_transcripts_that_need_deanonymisation}
\alias{get_transcripts_that_need_deanonymisation}
\title{Gets transcripts that have an anonymized summary but no deanonymized summary}
\usage{
get_transcripts_that_need_deanonymisation(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
A dataframe of transcripts to process
}
\description{
Gets transcripts that have an anonymized summary but no deanonymized summary
}
\keyword{internal}
