% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_summary.R
\name{get_transcripts_without_summary}
\alias{get_transcripts_without_summary}
\title{Gets all transcripts without a summary from the database}
\usage{
get_transcripts_without_summary(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame with transcripts that do not have a summary
}
\description{
Gets all transcripts without a summary from the database
}
\keyword{internal}
