% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_scoped_graph_helpers.R
\name{graph_get}
\alias{graph_get}
\title{Einzelner Graph-GET}
\usage{
graph_get(url, token, query = NULL)
}
\arguments{
\item{url}{Voll-URL. @param token Provider-Closure oder String. @param query optionale Query-Liste.}
}
\value{
list(status, content)
}
\description{
Einzelner Graph-GET
}
