% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{id_anonymize_locations}
\alias{id_anonymize_locations}
\title{Gives anonymized IDs to LOCATION tokens using NER_LOCATION_\%06d format}
\usage{
id_anonymize_locations(con, tokens)
}
\arguments{
\item{con}{Database connection}

\item{tokens}{A dataframe with a \code{token} column and \code{NER} column}
}
\value{
The updated tokens dataframe with NER values anonymized
}
\description{
Gives anonymized IDs to LOCATION tokens using NER_LOCATION_\%06d format
}
\keyword{internal}
