% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msgraph_func.R
\docType{data}
\name{INTERNAL_EMAIL_PATTERNS}
\alias{INTERNAL_EMAIL_PATTERNS}
\title{Internal Email Patterns}
\format{
An object of class \code{character} of length 11.
}
\usage{
INTERNAL_EMAIL_PATTERNS
}
\description{
Vector of email patterns considered internal (company emails).
Patterns are matched anywhere in the email address.
Dots are automatically escaped for regex matching.
}
\examples{
# Domain patterns
INTERNAL_EMAIL_PATTERNS <- c("@studyflix.de", "@tenhil.de")

# Exchange system addresses
INTERNAL_EMAIL_PATTERNS <- c("@studyflix.de", "/o=exchangelabs")

}
\keyword{datasets}
