% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_task_meeting_parser.R
\name{is_external_tool}
\alias{is_external_tool}
\title{Ist das Tool ein externes (nicht-Teams) VC-Tool?}
\usage{
is_external_tool(tool)
}
\arguments{
\item{tool}{Character(-Vektor) aus \code{\link[=extract_meeting_tool]{extract_meeting_tool()}}.}
}
\value{
Logical(-Vektor).
}
\description{
Ist das Tool ein externes (nicht-Teams) VC-Tool?
}
