% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_crm_export.R
\name{mark_transcripts_as_exported}
\alias{mark_transcripts_as_exported}
\title{Mark Transcripts as Exported to CRM}
\usage{
mark_transcripts_as_exported(con, transcript_ids)
}
\arguments{
\item{con}{Database connection}

\item{transcript_ids}{Vector of transcript IDs that were successfully exported}
}
\value{
Number of records updated
}
\description{
Updates the database to mark successfully exported transcripts
}
\keyword{internal}
