% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{merge_v_tags}
\alias{merge_v_tags}
\title{Merges speaker tags (< v Name >) into single tokens}
\usage{
merge_v_tags(df, token_col = "token")
}
\arguments{
\item{df}{A dataframe with tokens}

\item{token_col}{The name of the token column}
}
\value{
A dataframe with merged speaker tags
}
\description{
Merges speaker tags (< v Name >) into single tokens
}
\keyword{internal}
