% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_delegated_auth.R
\name{msgraph_delegated_store_read}
\alias{msgraph_delegated_store_read}
\title{Verschluesselten Token-Store lesen}
\usage{
msgraph_delegated_store_read(path, key)
}
\arguments{
\item{path}{Pfad zur verschluesselten Store-Datei.}

\item{key}{Entschluesselungsschluessel.}
}
\value{
Liste mit refresh_token, obtained_at, last_refreshed_at, tenant_id,
client_id, scopes.
}
\description{
Verschluesselten Token-Store lesen
}
