% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_map_call_event.R
\name{msgraph_map_calls_events}
\alias{msgraph_map_calls_events}
\title{Map calls with events and update mapping table.}
\usage{
msgraph_map_calls_events(con)
}
\arguments{
\item{con}{A PostgreSQL connection object.}
}
\value{
No return value. Updates the database table.
}
\description{
This function matches MSGraph call and event records, classifies events, and updates the mapping.msgraph_call_event table.
}
\examples{
msgraph_map_calls_events(con)
}
