% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{process_transcript_placeholders}
\alias{process_transcript_placeholders}
\title{Processes transcript placeholders}
\usage{
process_transcript_placeholders(
  con,
  known_information,
  tokens,
  content_id,
  scope
)
}
\arguments{
\item{con}{Database connection}

\item{known_information}{Known information about participants}

\item{tokens}{Token dataframe}

\item{content_id}{Content ID}

\item{scope}{Scope of processing}
}
\value{
Updated tokens dataframe
}
\description{
Processes transcript placeholders
}
\keyword{internal}
