% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{replace_emails}
\alias{replace_emails}
\title{Replaces emails in the NER column with \if{html}{\out{<EMAIL>}}}
\usage{
replace_emails(tokens_df)
}
\arguments{
\item{tokens_df}{A dataframe with tokens}
}
\value{
A dataframe with replaced emails
}
\description{
Replaces emails in the NER column with \if{html}{\out{<EMAIL>}}
}
\keyword{internal}
