% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{replace_phone_numbers}
\alias{replace_phone_numbers}
\title{Replaces phone numbers in the NER column with \if{html}{\out{<TEL>}}}
\usage{
replace_phone_numbers(tokenized_df)
}
\arguments{
\item{tokenized_df}{A dataframe with tokens}
}
\value{
A dataframe with replaced phone numbers
}
\description{
Replaces phone numbers in the NER column with \if{html}{\out{<TEL>}}
}
\keyword{internal}
