% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{replace_websites}
\alias{replace_websites}
\title{Replaces websites in the NER column with \if{html}{\out{<WEBSITE>}}}
\usage{
replace_websites(tokens_df)
}
\arguments{
\item{tokens_df}{A dataframe with tokens}
}
\value{
A dataframe with replaced websites
}
\description{
Replaces websites in the NER column with \if{html}{\out{<WEBSITE>}}
}
\keyword{internal}
