% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{retrieve_transcript_metadata}
\alias{retrieve_transcript_metadata}
\title{Retrieve Transcript Metadata}
\usage{
retrieve_transcript_metadata(access_token, user_id, start_date, end_date)
}
\description{
Gets transcript metadata from MS Graph API
}
\keyword{internal}
