% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_summary.R
\name{retry_llm}
\alias{retry_llm}
\title{Retries the llm call in case of transient errors}
\usage{
retry_llm(prompt, chat, max_attempts = 4, wait_base = 2)
}
\arguments{
\item{prompt}{A string containing the call transcript to be summarized.}

\item{chat}{An ellmer chat object}

\item{max_attempts}{The maximum number of retry attempts.}

\item{wait_base}{The base wait time for exponential backoff.}
}
\value{
summary of the text
}
\description{
Retries the llm call in case of transient errors
}
\keyword{internal}
