% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{safe_extract}
\alias{safe_extract}
\title{Safely extracts a column from a dataframe}
\usage{
safe_extract(df, col_name)
}
\arguments{
\item{df}{A dataframe}

\item{col_name}{The name of the column to extract}
}
\value{
The value of the column or NA if not found
}
\description{
Safely extracts a column from a dataframe
}
\keyword{internal}
