% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_scoped_transcripts.R
\name{vtt_to_plaintext}
\alias{vtt_to_plaintext}
\title{VTT-Transkript in Plaintext (Sprecher + Text, ohne Zeitstempel)}
\usage{
vtt_to_plaintext(vtt)
}
\arguments{
\item{vtt}{VTT-String.}
}
\value{
Plaintext-String.
}
\description{
VTT-Transkript in Plaintext (Sprecher + Text, ohne Zeitstempel)
}
