% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{anonymize_speaker}
\alias{anonymize_speaker}
\title{Anonymizes speaker tags like <v Max Müller> to SPEAKER/00001 in the NER column}
\usage{
anonymize_speaker(con, tokens)
}
\arguments{
\item{con}{Database connection}

\item{tokens}{A data frame with a column \code{token} and \code{NER}}
}
\value{
Updated tokens with speaker names anonymized in the NER column
}
\description{
Anonymizes speaker tags like <v Max Müller> to SPEAKER/00001 in the NER column
}
\keyword{internal}
