% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{fetch_transcripts_for_all_users}
\alias{fetch_transcripts_for_all_users}
\title{Fetch Transcripts for All Users}
\usage{
fetch_transcripts_for_all_users(
  users,
  all_events_categorised_df,
  calls,
  authentication_msgraph,
  start_date,
  end_date,
  existing_transcript_ids = character(0)
)
}
\description{
Iterates through all users and retrieves their transcripts
}
\keyword{internal}
