% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_load.R
\name{get_content_transcript_url}
\alias{get_content_transcript_url}
\title{Get Content from Transcript URL}
\usage{
get_content_transcript_url(access_token, content_url)
}
\description{
Downloads transcript content (VTT format)
}
\keyword{internal}
