% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_transcripts_anonymization.R
\name{get_latest_corenlp_path}
\alias{get_latest_corenlp_path}
\title{Get Latest CoreNLP Path}
\usage{
get_latest_corenlp_path()
}
\value{
Character string with path to CoreNLP directory
}
\description{
Finds the most recent CoreNLP installation
}
\keyword{internal}
