% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_scoped_graph_helpers.R
\name{graph_collect}
\alias{graph_collect}
\title{Graph-GET mit Paging (@odata.nextLink)}
\usage{
graph_collect(url, token, query = NULL)
}
\arguments{
\item{url}{Voll-URL. @param token Provider-Closure oder String. @param query optionale Query-Liste.}
}
\value{
list(status, error, value)
}
\description{
Graph-GET mit Paging (@odata.nextLink)
}
