% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msgraph_func.R
\name{load_preserved_names}
\alias{load_preserved_names}
\title{Load Preserved Company Names}
\usage{
load_preserved_names(
  file_path = "../../base-data/msgraph/text_for_summary_prompt/wettbewerber.txt"
)
}
\arguments{
\item{file_path}{Path to file with company names}
}
\value{
Character vector of company names
}
\description{
Loads company/brand names to preserve during anonymization
}
\keyword{internal}
