% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_calls.R
\name{msgraph_update_calls}
\alias{msgraph_update_calls}
\title{Retrieve Call Records from MSGraph and Update Database}
\usage{
msgraph_update_calls(con, access_token)
}
\arguments{
\item{con}{A PostgreSQL database connection object.}

\item{access_token}{MSGraph API access token.}
}
\value{
No return value. Updates database tables with new call records and participants.
}
\description{
Retrieves new call records from MSGraph since the last update, processes them, and updates relevant database tables.
}
\examples{
msgraph_update_calls(con, access_token)
}
