% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_events.R
\name{msgraph_update_events}
\alias{msgraph_update_events}
\title{Retrieve and Update Calendar Events from MSGraph}
\usage{
msgraph_update_events(con, access_token, startDate, user_id = NULL)
}
\arguments{
\item{con}{A PostgreSQL database connection object.}

\item{access_token}{MSGraph API access token.}

\item{startDate}{Date from which to retrieve calendar events.}

\item{user_id}{Optional. Specific user ID (intern ID) to filter calendar events.}
}
\value{
No return value. Updates database tables with new calendar events and participants.
}
\description{
Retrieves calendar events for users from MSGraph since a given start date, processes them, and updates relevant database tables.
}
\examples{
msgraph_update_events(con, access_token, startDate)
msgraph_update_events(con, access_token, startDate, user_id = 23)
}
