% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_users.R
\name{msgraph_update_users}
\alias{msgraph_update_users}
\title{Retrieve and upsert all users from MS Graph to PostgreSQL}
\usage{
msgraph_update_users(con, access_token)
}
\arguments{
\item{con}{PostgreSQL connection object}

\item{access_token}{MS Graph access token}
}
\value{
NULL. Upserts user data into the database.
}
\description{
Retrieve and upsert all users from MS Graph to PostgreSQL
}
\examples{
msgraph_update_users(access_token, con)
}
