% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgraph_scoped_graph_helpers.R
\name{resolve_user_id}
\alias{resolve_user_id}
\title{UPN -> ObjectId (List-Endpoint mit $filter)}
\usage{
resolve_user_id(upn, token)
}
\arguments{
\item{upn}{userPrincipalName. @param token Provider/String.}
}
\value{
list(status, id)
}
\description{
UPN -> ObjectId (List-Endpoint mit $filter)
}
